fix: Graceful shutdown for queue & playwright
The previous implementation was leaving open the pipe through the
playwright-go library and playwright's nodejs driver. This addresses the
issue by moving the playwright initialization outside the goroutine so
that it will be closed.
This commit also removes the useless q.Quit channel in favor of passing
a context instance for shutdown.