feat: Added an alternative temp dir
This is meant as a perf upgrade that will be tested and, in theory,
should increase performance by using a temp dir placed in the
executable's directory. This is meant for users that place this app in
a drive/partition that's not the system's. Will require more proofing
and maybe the temp dir path will be changed to be one inside the
downloads dir specified in the settings.